// src/models/music.model.ts
export class songs {
    album: string;
    artist: string;
    duration: number;
    image: string;
    name: string;
    song: string;

    constructor() {}
}

export class albums {
    artist: string;
    date: string;
    image: string;
    length: string;
    name: string;
    tracks: string;

    constructor() {}
}

export class artist {
    image: string;
    name: string;
    songs: string;

    constructor() {}
}